add role uc
