+++ apps/server/src/modules/user/repo/role.repo.ts
@@ -1,22 +1,18 @@
 import { EntityManager } from '@mikro-orm/mongodb';
 import { Injectable } from '@nestjs/common';
 import { EntityId } from '@shared/domain';
-import { Role, IPermissionsAndRoles } from '../entity';
+import { Role } from '../entity';
 
 interface CacheEntry {
 	[name: string]: { value: Role; validUntil: number };
 }
 
-// TODO: no query handled at the moment
-// TODO: write test and move
-// TODO: replace with generall solution if it is added
+// TODO: replace with https://mikro-orm.io/docs/caching/
 export class RoleCache {
 	cache: CacheEntry;
 
 	name: string;
 
-	// we do not registred changes by calling the old endpoint
-	// also the clear interval should short
 	clearInterval: 60000;
 
 	constructor(name = '') {
@@ -33,7 +29,7 @@ export class RoleCache {
 	add(selector: string, role: Role): void {
 		this.cache[selector] = {
 			value: role,
-			validUntil: Date.now() + this.clearInterval, // TODO: test if it is work
+			validUntil: Date.now() + this.clearInterval,
 		};
 	}
 
@@ -76,7 +72,7 @@ export class RoleRepo {
 		return role;
 	}
 
-	async resolvePermissionsById(id: EntityId): Promise<Role> {
+	async resolvePermissionsFromSubRolesById(id: EntityId): Promise<Role> {
 		if (this.cl2.get(id)) {
 			return this.cl2.get(id) as Role;
 		}
@@ -86,7 +82,7 @@ export class RoleRepo {
 
 		for (let i = 0; i < role.roles.length; i += 1) {
 			// eslint-disable-next-line no-await-in-loop
-			const resolvedSubRole = await this.resolvePermissionsById(role.roles[i]);
+			const resolvedSubRole = await this.resolvePermissionsFromSubRolesById(role.roles[i]);
 			permissions = [...permissions, ...resolvedSubRole.permissions];
 		}
 
@@ -94,17 +90,5 @@ export class RoleRepo {
 		role.permissions = uniquePermissions;
 		this.cl2.add(id, role);
 		return role;
-	}
-
-	async resolvePermissionsByIdList(ids: EntityId[]): Promise<IPermissionsAndRoles> {
-		const roles = await Promise.all(ids.map((id) => this.resolvePermissionsById(id)));
-
-		let permissions: string[] = [];
-		roles.forEach((role) => {
-			permissions = [...permissions, ...role.permissions];
-		});
-		permissions = [...new Set(permissions)];
-
-		return { roles, permissions };
 	}
 }
