+++ apps/server/src/modules/user/uc/index.ts
@@ -1,1 +1,2 @@
 export * from './user.uc';
+export * from './role.uc';
