+++ apps/server/src/modules/user/uc/user.uc.ts
@@ -2,18 +2,17 @@ import { Injectable } from '@nestjs/common';
 import { ICurrentUser } from '../../authentication/interface/jwt-payload';
 import { ResolvedUserMapper } from '../mapper';
 import { ResolvedUser } from '../controller/dto/ResolvedUser.dto';
+import { RoleUC } from './role.uc';
+import { UserRepo } from '../repo';
 
-import { UserRepo, RoleRepo } from '../repo';
-
-// TODO: IUserWithPermission as return
 @Injectable()
 export class UserUC {
-	constructor(private userRepo: UserRepo, private roleRepo: RoleRepo) {}
+	constructor(private readonly userRepo: UserRepo, private readonly roleUC: RoleUC) {}
 
 	async getUserWithPermissions(currentUser: ICurrentUser): Promise<ResolvedUser> {
 		const [user, resolved] = await Promise.all([
 			this.userRepo.findById(currentUser.userId),
-			this.roleRepo.resolvePermissionsByIdList(currentUser.roles),
+			this.roleUC.resolvePermissionsByIdList(currentUser.roles),
 		]);
 
 		const resolvedUser = ResolvedUserMapper.mapToResponse(user, resolved.permissions, resolved.roles);
